+++ src/controllers/commentController.js
@@ -2,6 +2,7 @@ import * as commentService from "../services/commentService.js";
 import {
   ValidationError,
   NotFoundError,
+  ForbiddenError,
 } from "../middlewares/errorMiddleware.js";
 
 export const createComment = async (req, res) => {
@@ -24,17 +25,22 @@ export const getComments = async (req, res) => {
 };
 
 export const updateComment = async (req, res) => {
+  const comment = await commentService.getCommentById(req.params.id);
+  if (comment.userId !== req.user.id) {
+    throw new ForbiddenError("댓글을 수정할 권한이 없습니다.");
+  }
   const updatedComment = await commentService.updateComment(
     req.params.id,
     req.body.content
   );
-  if (!updatedComment) {
-    throw new NotFoundError("댓글을 찾을 수 없습니다.");
-  }
   res.json(updatedComment);
 };
 
 export const deleteComment = async (req, res) => {
+  const comment = await commentService.getCommentById(req.params.id);
+  if (comment.userId !== req.user.id) {
+    throw new ForbiddenError("댓글을 삭제할 권한이 없습니다.");
+  }
   await commentService.deleteComment(req.params.id);
   res.status(204).send();
 };
